+++ src/api/banner/index.ts
@@ -1,0 +1,11 @@
+import { requestJWT } from "../../utils/axiosConfig";
+import { ResponseBanner } from "./type";
+
+export async function getBanner() {
+  try {
+    const res: ResponseBanner = await requestJWT.get("/banner");
+    return res;
+  } catch (err) {
+    console.log(err);
+  }
+}
