+++ src/api/product/type.ts
@@ -1,0 +1,9 @@
+import { ApiResponseType, ProductsType } from "types";
+
+export interface ResProductType extends ApiResponseType {
+  data: {
+    status: boolean;
+    data: ProductsType[];
+    totalPage: number;
+  };
+}
