+++ src/api/refreshToken/type.ts
@@ -1,0 +1,6 @@
+import { CurrentAccount } from "types";
+
+export interface ResResfreshType {
+  status: boolean;
+  data: CurrentAccount["refresh"];
+}
