+++ src/api/search/type.ts
@@ -1,0 +1,8 @@
+import { ApiResponseType, ProductsType } from "types";
+
+export interface ResSearchType extends ApiResponseType {
+  data: {
+    status: boolean;
+    data: ProductsType[];
+  };
+}
