--- src/services/productNewsService.ts
@@ -1,10 +1,0 @@
-import { requestJWT } from "../utils/axiosConfig";
-
-export async function getProductNew() {
-  try {
-    const res = await requestJWT.get(`product/new_product`);
-    return res;
-  } catch (err) {
-    console.log(err);
-  }
-}
