+++ src/types/api.type.ts
@@ -1,0 +1,31 @@
+export interface ApiResponseType {
+  status: number;
+  statusText: string;
+  headers: {
+    "content-length": string;
+    "content-type": string;
+  };
+  config: {
+    transitional: {
+      silentJSONParsing: boolean;
+      forcedJSONParsing: boolean;
+      clarifyTimeoutError: boolean;
+    };
+    adapter: string[];
+
+    timeout: number;
+    xsrfCookieName: string;
+    xsrfHeaderName: string;
+    maxContentLength: number;
+    maxBodyLength: number;
+    env: object;
+    headers: {
+      Accept: string;
+      token: string;
+    };
+    baseURL: string;
+    method: string;
+    url: string;
+  };
+  request: object;
+}
