+++ src/types/formEvent.type.ts
@@ -1,0 +1,2 @@
+export type InputEvent = React.ChangeEvent<HTMLInputElement>;
+export type FormEvent = React.FormEvent<HTMLFormElement>;
