+++ src/types/input.type.ts
@@ -1,0 +1,7 @@
+export interface InputProps {
+  placeholder: string;
+  typeInput: string;
+  valueBtn: string;
+  className: string;
+  onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+}
